Block and connect to database instead of using a thread
When you delay the connection to a thread it will be delayed up until
players join the server. 
By this time all the configs will have executed.
We have to establish the connection early so the cvars we generate
at runtime exist by the time the server configs are executed